Refactor builder UI into package
