+++ builder_ui/__init__.py
@@ -1,0 +1,6 @@
+"""UI components and tools for building particle scenes."""
+
+from .fields import SliderField, ColorField, KeyField
+from .sidebar import SidebarUI
+
+__all__ = ["SliderField", "ColorField", "KeyField", "SidebarUI"]
